Refactor: replace MovieItem component with a new implementation in the list directory, enhancing functionality and styling
