+++ src/screens/mymovies/list/components/MovieItem/index.styled.ts
